move compute.go.
