+++ cmd/compute/instance/instance.go
@@ -1,7 +1,7 @@
 package instance
 
 import (
-	exocmd "github.com/exoscale/cli/cmd"
+	"github.com/exoscale/cli/cmd/compute"
 	"github.com/spf13/cobra"
 )
 
@@ -34,5 +34,5 @@ var instanceCmd = &cobra.Command{
 }
 
 func init() {
-	exocmd.ComputeCmd.AddCommand(instanceCmd)
+	compute.ComputeCmd.AddCommand(instanceCmd)
 }
