+++ cmd/compute/private_network/private_network.go
@@ -1,7 +1,7 @@
 package private_network
 
 import (
-	exocmd "github.com/exoscale/cli/cmd"
+	"github.com/exoscale/cli/cmd/compute"
 	"github.com/spf13/cobra"
 )
 
@@ -12,5 +12,5 @@ var privateNetworkCmd = &cobra.Command{
 }
 
 func init() {
-	exocmd.ComputeCmd.AddCommand(privateNetworkCmd)
+	compute.ComputeCmd.AddCommand(privateNetworkCmd)
 }
