+++ cmd/compute/sks/sks.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	exocmd "github.com/exoscale/cli/cmd"
+	"github.com/exoscale/cli/cmd/compute"
 	"github.com/exoscale/cli/pkg/globalstate"
 )
 
@@ -20,5 +20,5 @@ var sksCmd = &cobra.Command{
 }
 
 func init() {
-	exocmd.ComputeCmd.AddCommand(sksCmd)
+	compute.ComputeCmd.AddCommand(sksCmd)
 }
