+++ cmd/compute/ssh_key/ssh_key.go
@@ -1,7 +1,7 @@
 package ssh_key
 
 import (
-	exocmd "github.com/exoscale/cli/cmd"
+	"github.com/exoscale/cli/cmd/compute"
 	"github.com/spf13/cobra"
 )
 
@@ -11,5 +11,5 @@ var computeSSHKeyCmd = &cobra.Command{
 }
 
 func init() {
-	exocmd.ComputeCmd.AddCommand(computeSSHKeyCmd)
+	compute.ComputeCmd.AddCommand(computeSSHKeyCmd)
 }
